Add CallDialog component for incoming calls
Added a new CallDialog component for displaying incoming call notifications. This component includes a modal dialog that appears when an incoming call is detected, providing options to answer or decline the call. The dialog can also play and pause an audio notification depending on visibility status. Subsequently, the CallDialog component is imported into the ProfileHeader component.